fix: 修复 NoFall 检测中的类型转换错误
- 将 player.getFallDistance() 返回的 float 类型显式转换为 Double
- 修复 Maven 编译失败问题 (incompatible types: float cannot be converted to java.lang.Double)

+++ src/main/java/nb114514/antibitch/AntiBitch.java
@@ -575,7 +575,7 @@ public class AntiBitch extends JavaPlugin implements Listener {
 
         // NoFall 检测（娱乐性质：检测无坠落伤害）
         if (!player.isOnGround() && !player.isFlying()) {
-            lastFallHeight.put(playerId, player.getFallDistance());
+            lastFallHeight.put(playerId, (double) player.getFallDistance());
         }
     }
